Revert "Add observatory Bonjour service to built iOS Info.plist bundle (#64988)"
This reverts commit 4fde217dea5b5404660a02ed4cc1d9071085098a.
